feat: various improvements
- use raw mode to avoid printing shortcut keys
- always bind shortcuts when `server.listen` is called
- move shortcut logic into new module: src/node/server/shortcuts.ts
